# contract/openapi.yaml
openapi: 3.1.0
info:
  title: MyBooking API
  version: 0.0.0
...
paths:
...
  /api/v1/experts/{publicId}/bookings:
    post:
      operationId: createBooking
      description: Гостевое доказательство погашается после успешного создания. Токен доступа отправляется только письмом.
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingCreated'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateInput'
...
components:
  schemas:
...
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Публичный код без раскрытия чужих сведений.
        message:
          type: string
          description: Безопасное описание на русском языке.
...
    BookingCreateInput:
      type: object
      required:
        - guestProof
        - guestName
        - guestTimezone
        - startAt
        - durationMinutes
        - topic
      properties:
        guestProof:
          type: string
        guestName:
          type: string
        guestTimezone:
          type: string
        startAt:
          $ref: '#/components/schemas/OffsetDateTime'
        durationMinutes:
          type: number
          enum:
            - 15
            - 30
            - 60
        topic:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 2000
      allOf:
        - $ref: '#/components/schemas/ConsentInput'
    BookingCreated:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - pending
      allOf:
        - $ref: '#/components/schemas/BookingDetail'
    BookingDetail:
      type: object
      required:
        - id
        - status
        - startAt
        - endAt
        - topic
        - guestName
        - guestTimezone
        - expertName
        - expertPublicId
        - completed
      properties:
        id:
          type: string
        status:
          type: string
          enum:
            - pending
            - confirmed
            - rejected
            - withdrawn
            - expired
            - cancelled
        startAt:
          $ref: '#/components/schemas/OffsetDateTime'
        endAt:
          $ref: '#/components/schemas/OffsetDateTime'
        topic:
          type: string
        description:
          type: string
        guestName:
          type: string
        guestTimezone:
          type: string
        expertName:
          type: string
        expertPublicId:
          type: string
        closedReason:
          type: string
          enum:
            - manual
            - conflict
            - schedule_changed
            - deadline
        completed:
          type: boolean
...
    ConsentInput:
      type: object
      required:
        - consentVersion
        - consentAccepted
      properties:
        consentVersion:
          type: string
        consentAccepted:
          type: boolean
          enum:
            - true
...
    OffsetDateTime:
      type: string
      format: date-time
      description: Момент ISO 8601 с явным смещением; сервер определяет однозначный UTC-момент.
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: cookie
      name: mybooking_session
    BearerAuth:
      type: http
      scheme: Bearer